Beyblade!

+++ src/main/java/com/kuuhaku/model/Beyblade.java
@@ -120,6 +120,6 @@ public class Beyblade {
     }
 
     public float getKDA() {
-        return wins / (loses == 0 ? 1 : loses);
+        return (float) wins / (loses == 0 ? 1 : loses);
     }
 }
